feat(secure dns): add DNS server helper methods in DNSProtocolType.swift
