+++ IVPNClient/Utilities/Extensions/NETunnelProviderProtocol+Ext.swift
@@ -52,13 +52,13 @@ extension NETunnelProviderProtocol {
         if let dnsServers = openVPNdnsServers(), dnsServers != [""] {
             sessionBuilder.dnsServers = dnsServers
             
-            switch SecureDNS.preferredDNSProtocol() {
+            switch DNSProtocolType.preferred() {
             case .doh:
                 sessionBuilder.dnsProtocol = .https
-                sessionBuilder.dnsHTTPSURL = URL.init(string: SecureDNS.getServerURL(address: UserDefaults.shared.customDNS) ?? "")
+                sessionBuilder.dnsHTTPSURL = URL.init(string: DNSProtocolType.getServerURL(address: UserDefaults.shared.customDNS) ?? "")
             case .dot:
                 sessionBuilder.dnsProtocol = .tls
-                sessionBuilder.dnsTLSServerName = SecureDNS.getServerName(address: UserDefaults.shared.customDNS)
+                sessionBuilder.dnsTLSServerName = DNSProtocolType.getServerName(address: UserDefaults.shared.customDNS)
             default:
                 sessionBuilder.dnsProtocol = .plain
             }
